Get k8s version EOL date from global config instead of global state
